Rename the shared folder to avoid a conflict with reserved keywords
