+++ src/DoomerPublish/DoomerPublish.lib/Tools/AcsParseService/AcsContext.cs
@@ -1,4 +1,4 @@
-﻿using DoomerPublish.Tools.Shared;
+﻿using DoomerPublish.Tools.Common;
 
 namespace DoomerPublish.Tools.Acs;
