+++ src/DoomerPublish/DoomerPublish.lib/Tools/DecorateParseService/DecorateContext.cs
@@ -1,5 +1,5 @@
 ﻿using DoomerPublish.Tools.Acs;
-using DoomerPublish.Tools.Shared;
+using DoomerPublish.Tools.Common;
 
 namespace DoomerPublish.Tools.Decorate;
